Debloat copy.go
