+++ cmd/copy.go
@@ -32,7 +32,6 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"time"
 
@@ -944,56 +943,12 @@ func (cca *cookedCopyCmdArgs) processCopyJobPartOrders() (err error) {
 	case common.EFromTo.LocalBlob(),
 		common.EFromTo.LocalBlobFS(),
 		common.EFromTo.LocalFile():
-		var traverser resourceTraverser
-
-		dst := cca.destination
-
-		if cca.destinationSAS != "" {
-			destURL, err := url.Parse(dst)
-
-			if err != nil {
-				return err
-			}
-
-			destURL = copyHandlerUtil{}.appendQueryParamToUrl(destURL, cca.destinationSAS)
-			dst = destURL.String()
-		}
-
-		traverser, err = initResourceTraverser(cca.source, cca.fromTo.From(), nil, nil, &cca.followSymlinks, cca.listOfFilesChannel, cca.recursive, func() {})
+		var e *copyEnumerator
+		e, err = cca.initEnumerator(jobPartOrder, ctx)
 		if err != nil {
 			return err
 		}
 
-		isDir := traverser.isDirectory(false)
-		isDestDir := cca.isDestDirectory(dst, &ctx)
-
-		if isDir && !cca.recursive {
-			return errors.New("cannot copy from container or directory without --recursive or trailing wildcard (/*)")
-		}
-
-		filters := cca.initModularFilters()
-		processor := func(object storedObject) error {
-			src := cca.makeEscapedRelativePath(true, isDestDir, object)
-			dst := cca.makeEscapedRelativePath(false, isDestDir, object)
-
-			transfer := common.CopyTransfer{
-				Source:           src,
-				Destination:      dst,
-				LastModifiedTime: object.lastModifiedTime,
-				SourceSize:       object.size,
-			}
-
-			return addTransfer(&jobPartOrder, transfer, cca)
-		}
-		finalizer := func() error {
-			if len(jobPartOrder.Transfers) == 0 {
-				return errors.New("cannot find source to upload")
-			}
-
-			return dispatchFinalPart(&jobPartOrder, cca)
-		}
-
-		e := newCopyEnumerator(traverser, filters, processor, finalizer)
 		err = e.enumerate()
 	case common.EFromTo.BlobLocal():
 		e := copyDownloadBlobEnumerator(jobPartOrder)
@@ -1058,163 +1013,6 @@ func (cca *cookedCopyCmdArgs) processCopyJobPartOrders() (err error) {
 	}
 
 	return nil
-}
-
-func initPipeline(ctx context.Context, location common.Location, credential common.CredentialInfo) (p pipeline.Pipeline, err error) {
-	switch location {
-	case common.ELocation.Local():
-		return nil, nil
-	case common.ELocation.Blob():
-		p, err = createBlobPipeline(ctx, credential)
-	case common.ELocation.File():
-		p, err = createFilePipeline(ctx, credential)
-	case common.ELocation.BlobFS():
-		p, err = createBlobFSPipeline(ctx, credential)
-	default:
-		err = fmt.Errorf("can't produce new pipeline for location %s", location)
-	}
-
-	return
-}
-
-func getPathBeforeFirstWildcard(path string) string {
-	if strings.Index(path, "*") == -1 {
-		return path
-	}
-
-	firstWCIndex := strings.Index(path, "*")
-	result := replacePathSeparators(path[:firstWCIndex])
-	lastSepIndex := strings.LastIndex(result, "/")
-	result = result[:lastSepIndex+1]
-
-	return result
-}
-
-func (cca *cookedCopyCmdArgs) makeEscapedRelativePath(source bool, dstIsDir bool, object storedObject) (relativePath string) {
-	var pathEncodeRules = func(path string) string {
-		loc := common.ELocation.Unknown()
-
-		if source {
-			loc = cca.fromTo.From()
-		} else {
-			loc = cca.fromTo.To()
-		}
-		pathParts := strings.Split(path, common.AZCOPY_PATH_SEPARATOR_STRING)
-
-		// Encode disallowed characters on windows explicit to downloads.
-		// TODO: Inform tests of this
-		// TODO: on upload refactor, reverse this operation.
-		if loc == common.ELocation.Local() && !source && runtime.GOOS == "windows" {
-			invalidChars := `<>\/:"|?*` + string(0x00)
-
-			for _, c := range strings.Split(invalidChars, "") {
-				for k, p := range pathParts {
-					pathParts[k] = strings.ReplaceAll(p, c, url.PathEscape(c))
-				}
-			}
-		} else if loc != common.ELocation.Local() {
-			for k, p := range pathParts {
-				pathParts[k] = url.PathEscape(p)
-			}
-		}
-
-		path = strings.Join(pathParts, "/")
-		return path
-	}
-
-	// source is a EXACT path to the file.
-	if object.relativePath == "" {
-		// If we're finding an object from the source, it returns "" if it's already got it.
-		// If we're finding an object on the destination and we get "", we need to hand it the object name (if it's pointing to a folder)
-		if source {
-			relativePath = ""
-		} else {
-			if dstIsDir {
-				relativePath = "/" + object.name
-			} else {
-				relativePath = ""
-			}
-		}
-
-		return pathEncodeRules(relativePath)
-	}
-
-	// If it's out here, the object is contained in a folder, or was found via a wildcard.
-
-	relativePath = "/" + strings.Replace(object.relativePath, common.OS_PATH_SEPARATOR, common.AZCOPY_PATH_SEPARATOR_STRING, -1)
-
-	if !source && !pathPointsToContents(cca.source) {
-		// We ONLY need to do this adjustment to the destination.
-		// The source SAS has already been removed. No need to convert it to a URL or whatever.
-		// Save to a directory
-		relativePath = "/" + filepath.Base(cca.source) + relativePath
-	}
-
-	return pathEncodeRules(relativePath)
-}
-
-// TODO: Find a better place for me and my friends (makeEscapedRelativePath, getPathBeforeFirstWildcard, initPipeline, etc.)
-// In local cases, many wildcards may be used, hence string.Contains
-// In non-local cases, only a trailing wildcard may be used: ex. https://myAccount.blob.core.windows.net/container/*
-// In both cases, we want to copy the contents of the matches to the exact path specified on the destination.
-// Without this, a directory is created at the destination, and everything is placed under it.
-func pathPointsToContents(path string) bool {
-	return strings.Contains(path, "*")
-}
-
-// This is condensed down into an individual function as we don't end up re-using the destination traverser at all.
-// This is just for the directory check.
-func (cca *cookedCopyCmdArgs) isDestDirectory(dst string, ctx *context.Context) bool {
-	destTraverser, err := initResourceTraverser(dst, cca.fromTo.To(), ctx, &cca.credentialInfo, &cca.followSymlinks, nil, cca.recursive, func() {})
-	if err != nil {
-		return false
-	}
-
-	return destTraverser.isDirectory(true)
-}
-
-// Initialize the modular filters outside of copy to increase readability.
-func (cca *cookedCopyCmdArgs) initModularFilters() []objectFilter {
-	filters := make([]objectFilter, 0) // same as []objectFilter{} under the hood
-
-	if len(cca.includePatterns) != 0 {
-		filters = append(filters, &includeFilter{patterns: cca.includePatterns})
-	}
-
-	if len(cca.excludePatterns) != 0 {
-		for _, v := range cca.excludePatterns {
-			filters = append(filters, &excludeFilter{pattern: v})
-		}
-	}
-
-	// include-path is not a filter, therefore it does not get handled here.
-	// Check up in cook() around the list-of-files implementation as include-path gets included in the same way.
-
-	if len(cca.excludePathPatterns) != 0 {
-		for _, v := range cca.excludePathPatterns {
-			filters = append(filters, &excludeFilter{pattern: v, targetsPath: true})
-		}
-	}
-
-	if len(cca.excludeBlobType) != 0 {
-		excludeSet := map[azblob.BlobType]bool{}
-
-		for _, v := range cca.excludeBlobType {
-			excludeSet[v] = true
-		}
-
-		filters = append(filters, &excludeBlobTypeFilter{blobTypes: excludeSet})
-	}
-
-	if len(cca.includeFileAttributes) != 0 {
-		filters = append(filters, buildAttrFilters(cca.includeFileAttributes, cca.source, true)...)
-	}
-
-	if len(cca.excludeFileAttributes) != 0 {
-		filters = append(filters, buildAttrFilters(cca.excludeFileAttributes, cca.source, false)...)
-	}
-
-	return filters
 }
 
 // wraps call to lifecycle manager to wait for the job to complete
